# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing Pointlings, users, items, and XP system",
        "title": "Pointlings API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/pointlings": {
            "post": {
                "description": "Create a new pointling for a user with optional nickname",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pointlings"
                ],
                "summary": "Create a new pointling",
                "parameters": [
                    {
                        "description": "Pointling creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createPointlingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Pointling"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing user_id",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pointlings/user/{user_id}": {
            "get": {
                "description": "Get all pointlings owned by a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pointlings"
                ],
                "summary": "List user's pointlings",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Pointling"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pointlings/{pointling_id}": {
            "get": {
                "description": "Get detailed information about a specific pointling",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pointlings"
                ],
                "summary": "Get pointling details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pointling ID",
                        "name": "pointling_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pointling"
                        }
                    },
                    "400": {
                        "description": "Invalid pointling ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pointling not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pointlings/{pointling_id}/nickname": {
            "patch": {
                "description": "Update or remove a pointling's nickname",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pointlings"
                ],
                "summary": "Update pointling nickname",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pointling ID",
                        "name": "pointling_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nickname update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateNicknameRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid pointling ID or request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pointling not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{user_id}/points": {
            "patch": {
                "description": "Set a new point balance for a user (cannot be negative)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user's point balance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Points update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updatePointsRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid user ID, request body, or negative balance",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.createPointlingRequest": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string",
                    "example": "MyPointling"
                },
                "user_id": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
...
        "handlers.updateNicknameRequest": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string",
                    "example": "CoolPointling"
                }
            }
        },
        "handlers.updatePointsRequest": {
            "type": "object",
            "properties": {
                "new_balance": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 1000
                }
            }
        },
...
        "models.JSONMap": {
            "type": "object",
            "additionalProperties": true
        },
...
        "models.Pointling": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_xp": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "look_json": {
                    "$ref": "#/definitions/models.JSONMap"
                },
                "nickname": {
                    "type": "string"
                },
                "personality_id": {
                    "type": "integer"
                },
                "pointling_id": {
                    "type": "integer"
                },
                "required_xp": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
...
}